Sort out [Related] vs. [Params] vs. context hypotheses
We want the [monotonicity] tactic to use hypotheses from the context as
one source of relational properties. Until now, we did this by hooking
them in as [Related] instances. However, such instances are filtered by
any [Params] instance that has been declared for the head of the related
terms.

Instead, we now hook them in directly as [CandidateProperty] instances.
This way, the will always override the regular database, regardless of
whether a [Params] declaration exists and how specialized the hypothesis
from the context is. As a nice side-effect, in this new setting
[monotonicity] subsumes direct use of [Related] instances, so that we
can drop [related_rstep].