New automation tactic [rauto]
Unlike [solve_monotonic], the new tactic is embedded into the typeclass
resolution mechanism through the [RAuto] class. This enables
backtracking across steps, and using [RAuto] as a component for more
elaborate classes and tactics should be preferable to [solve_monotonic]
(for one thing the [Typeclasses eauto := debug] traces would be easier
to decypher).